requirements.txt
requests
python-dotenv
openai
uvicorn[standard]
replicate
